Refactor use internal Elevation type
The elevation requires specific magic values. The apparent values
are chosen to match the values in the Material Design 2 specifications,
but some magic values are set accordingly in the setElevation method
to provide the expected visual effect.

Using a specific type allows to abstract the special values while
providing extra documentation.