Refactor use event transitions instead of signal transitions
This involves significantly less boilerplate code, and is quite elegant.
The PushButton filters with events it relays to PushButtonState,
which can then keep the transition definitions simple.

Note: For some reason, one of the event transitions causes unexpected
behavior. I'm not sure why, or why that one in particular.
I left the signal transitionas a workaround. (I don't even understand
why one would work and not the other... it is a really surprising bug
to me.)