Fix 'pressed' state bug, drastically simplify event handling
Reworded the statechart to match transition names to Qt events

Main takeaways and explanation:

I learned that the state machine was receiving the button events
anyway, so it wasn't necessary for the button to be posting them.

In the case of key events, even though the button was only posting
those related to the 'space' key, the machine was receiving them all.
Because the transition was triggered on a generic KeyPress type,
the event originating from the TAB during FocusOut was triggering
the transition to the corresponding 'pressed" state.

Instead of filtering the event at the button, a transition can be
defined to be triggered on a key event related to a specific key
(e.g. 'space'). Once that's done, the butotn doesn't need to filter
events, nor to send them (the machine receives them anyway).
That simplifies considerable the API between the button and its state.

The state still holds on a reference to the button in order to define
the transitions, but it seems a lot clearer that the state doesn't
issue any commands to the button. Instead, the button subscribes
to the signals that the state transitions emit.